taskmaster : consider all field of config file

+++ taskmaster/src/process.js
@@ -19,6 +19,9 @@ module.exports = class Process {
       this.killSignal = config["stopsignal"] ? config["stopsignal"] : 'SIGTERM';
       this.env = config["env"] ? config["env"] : null;
 
+      this.startRetries = config["startretries"] ? config["startretries"] : 3;
+      this.tries = 0;
+      this.startTime = config["starttime"] ? config["starttime"] * 1000 : 1 * 1000;
       this.out = 1;
       this.err = 2;
 
@@ -31,46 +34,70 @@ module.exports = class Process {
 
     start()
     {
+        this.tries++;
         this.updateState("STARTING");
         startProcess(this, function(res) {
-            if (res.status === "RUNNING")
+            if (res.state === "RUNNING")
             {
+                
+                this.updateState("RUNNING");
                 this.pid = res.payload.pid;
                 this.ppid = res.payload.ppid;
-                this.updateState("RUNNING");
-            }
-            if (res.status === "EXITED")
+            } 
+            else if (res.state === "EXITED" || res.state === "BACKOFF")
             {
-                this.updateState("EXITED");
-                this.reload(res.payload);
+                this.updateState(res.state);
+                if (this.isReloadable(res.payload) === true)
+                    this.start();
             }
         }.bind(this));
     }
 
-    reload(payload)
+    isReloadable(payload)
     {
         let code = payload.code;
-        let signal = payload.signal;
-
-        switch (this.autorestart) {
-            case true:
-              console.log('Restart ' + this.name);
-              break;
-            case false:
-                break;
-            case "unexpected":
-                if (isExitCodesValid(this.exitcodes, code) === false)
-                    console.log(`Not OK :  ${code} is not part of ` + this.exitcodes + ' -> reboot');
-                else
+        
+        if (this.state === "BACKOFF")
+        {
+            if (this.tries === this.startRetries)
+            {
+                this.updateState("FATAL");
+                return false;
+            }
+            return true;
+        }
+        if (this.state === "EXITED")
+        {
+            switch (this.autorestart) {
+                case true:
+                  console.log('Restart ' + this.name);
+                  return true;
+                case false:
                     console.log(`exit name : ${this.name} wide code : ${code}`);
-              break;
-            default:
-                console.log(`exit name : ${this.name} wide code : ${code}`);
-          }
+                    return false;
+                case "unexpected":
+                    if (isExitCodesValid(this.exitcodes, code) === false)
+                    {
+                        console.log(`Not OK :  ${code} is not part of ` + this.exitcodes + ' -> reboot');
+                        return true;
+                    }
+                    else
+                    {
+                        console.log(`exit name : ${this.name} wide code : ${code}`);
+                        return false;
+                    }
+                default:
+                {
+                    console.log(`exit name : ${this.name} wide code : ${code}`);
+                    return false;   
+                }
+              }
+        }
     }
 
     updateState(state)
     {
+        console.log(`${this.name} changed state from ${this.state} to ${state}`);
         this.state = state;
         this.stateInt = getStateInt(state);
     }
